paymcp-video-demo/requirements.txt

mcp
paymcp>=0.4.2
aiohttp
python-dotenv
